refactor(cardano-services): embed program option defaults into their respective modules
