+++ packages/cardano-services/src/Program/index.ts
@@ -1,4 +1,3 @@
-export * from './defaults';
 export * from './errors';
 export * from './loadHttpServer';
 export * from './loadTxWorker';
