+++ packages/cardano-services/src/Program/loadTxWorker.ts
@@ -5,6 +5,10 @@ import { createDnsResolver } from './utils';
 import { createLogger } from 'bunyan';
 import { getOgmiosTxSubmitProvider, getRunningTxSubmitWorker } from './services';
 
+export const PARALLEL_MODE_DEFAULT = false;
+export const PARALLEL_TXS_DEFAULT = 3;
+export const POLLING_CYCLE_DEFAULT = 500;
+
 export enum TxWorkerOptionDescriptions {
   Parallel = 'Parallel mode',
   ParallelTxs = 'Parallel transactions',
